Changed project type to maven
