+++ src/main/resources/plugin.yml
@@ -1,0 +1,8 @@
+name: CapesM
+main: leonidm.capes.CapesMain
+version : 1.0.2
+api-version: 1.13
+commands:
+  cape:
+    usage: '/cape [set/remove/reload]'
+    description: 'Set banner from hand as a cape or remove it'
